validate address specific routes against jwt owner

+++ src/app/api/v1/retrofunding/rounds/[roundId]/ballots/[ballotCasterAddressOrEns]/impactMetrics/[impactMetricId]/route.ts
@@ -2,6 +2,8 @@ import { NextResponse, type NextRequest } from "next/server";
 import { authenticateApiUser } from "@/app/lib/auth/serverAuth";
 import { traceWithUserId } from "@/app/api/v1/apiUtils";
 import { deleteBallotMetric } from "@/app/api/common/ballots/updateBallot";
+import { isAddress } from "ethers";
+import { resolveENSName } from "@/app/lib/ENSUtils";
 
 export async function DELETE(
   request: NextRequest,
@@ -21,6 +23,16 @@ export async function DELETE(
 
   return await traceWithUserId(authResponse.userId as string, async () => {
     const { roundId, ballotCasterAddressOrEns, impactMetricId } = route.params;
+
+    const address = isAddress(ballotCasterAddressOrEns)
+      ? ballotCasterAddressOrEns.toLowerCase()
+      : await resolveENSName(ballotCasterAddressOrEns);
+
+    if (authResponse.userId?.toLowerCase() !== address) {
+      return new Response("Unauthorized to perform action on this address", {
+        status: 401,
+      });
+    }
 
     try {
       await deleteBallotMetric(
